Set user agent for end-to-end tests
This adds a distinctive user agent to the fetch for the end-to-end tests, so that logs are easier to navigate. It also doesn't try to build the target resource IRI anymore, and instead relies on the server sending it back with the creation response. A slug is still passed to the server to suggest a resource name.