Disable running tests on MacOS (#1785)
MacOS tests have always been very flaky, and lately the browser-based tests on MacOS have been consistently failing. We haven't managed to reproduce the issues locally, so we'll disable running the tests on macOS until we have time to investigate properly. This probably means only revisiting when there is an actual issue on MacOS, but the likelyhood of a NodeJS or browser issue to be MacOS-specific is low enough that this is an acceptable risk.