Set recursive VC access (#1832)
This introduces a new `options.inherit` flag on the `setVcAccess` function. If set to `true`, in addition to setting the VC access on the resource's own Access Control, it also sets it on its Member Access Control so that access is cascaded to contained resources. For backwards compatibility, `options.inherit` defaults to `false`.

This feature is added to be used in conjunction with recursive Access Grants. In the case of a non-recursive Access Grant, it is recommanded to leave the flag to the default `false` value, so that only the target resource has its access changed.