Universal API no longer throws on missing ACL
Fixes #1549

A resource having an ACL link, but this link not resolving to an ACL is
a legitimate situation in the ACL protocol. The universal API had a bug
that was not allowing the underlying ACL API to create the target ACL if
it was missing. It is now supported.